requirements.txt
streamlit
langchain
python-docx
google-generativeai
faiss-cpu
python-dotenv